mise a jour de la route post

+++ modules/cloudinary.js
@@ -3,8 +3,11 @@ const fs = require("fs")
 const uniqid = require("uniqid")
 
 async function upload(req, res, next) {
+  if (!req.files?.photoFromFront) {
+    next()
+  }
   const photoPath = `../tmp/${uniqid()}.jpg`
-  const resultMove = await req.files.photoFromFront.mv(photoPath)
+  const resultMove = await req.files?.photoFromFront.mv(photoPath)
   const resultCloudinary = await cloudinary.uploader.upload(photoPath)
   fs.unlinkSync(photoPath)
   req.file.cloudinary_url = resultCloudinary.secure_url
